refactor(api): move route to src
