+++ api/lib/routes.js
@@ -1,5 +1,4 @@
 import * as healthcheck from '../src/shared/application/healthcheck/index.js';
-import * as adminMembers from './application/admin-members/index.js';
 import * as authentication from './application/authentication/index.js';
 import * as cache from './application/cache/index.js';
 import * as campaignParticipations from './application/campaign-participations/index.js';
@@ -30,7 +29,6 @@ import * as userOrgaSettings from './application/user-orga-settings/index.js';
 import * as users from './application/users/index.js';
 
 const routes = [
-  adminMembers,
   authentication,
   cache,
   campaignParticipations,
